YUK-6284 New customer account and device enrollment import process YUKRV-502   New or Updated Stars Daos for BGE Project
- Fixes for review feedback
- Initial add of service helper layer

+++ yukon-client/stars/src/com/cannontech/stars/core/dao/impl/CallReportDaoImpl.java
@@ -2,28 +2,46 @@ package com.cannontech.stars.core.dao.impl;
 
 import java.util.List;
 
+import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.simple.SimpleJdbcTemplate;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.cannontech.common.util.ChunkingSqlTemplate;
+import com.cannontech.common.util.SqlGenerator;
 import com.cannontech.common.util.SqlStatementBuilder;
 import com.cannontech.database.IntegerRowMapper;
 import com.cannontech.stars.core.dao.CallReportDao;
 import com.cannontech.stars.core.dao.ECMappingDao;
 
-public class CallReportDaoImpl implements CallReportDao {
+public class CallReportDaoImpl implements CallReportDao, InitializingBean {
 
     private SimpleJdbcTemplate simpleJdbcTemplate;
     private ECMappingDao ecMappingDao;
+    private ChunkingSqlTemplate<Integer> chunkyJdbcTemplate;
     
     @Override
     @Transactional
     public void deleteByAccount(int accountId) {
         List<Integer> callReportIds = getByAccount(accountId);
-        ecMappingDao.deleteECToCallReportMapping(callReportIds);
-        SqlStatementBuilder sql = new SqlStatementBuilder();
-        sql.append("DELETE FROM CallReportBase WHERE CallId IN (", callReportIds, ")");
-        simpleJdbcTemplate.update(sql.toString());
+        if(!callReportIds.isEmpty()) {
+            ecMappingDao.deleteECToCallReportMapping(callReportIds);
+            chunkyJdbcTemplate.update(new CallReportBaseDeleteSqlGenerator(), callReportIds);
+        }
+    }
+    
+    /**
+     * Sql generator for deleting call reports from CallReportBase, useful for bulk deleteing
+     * with chunking sql template.
+     */
+    private class CallReportBaseDeleteSqlGenerator implements SqlGenerator<Integer> {
+
+        @Override
+        public String generate(List<Integer> callReportIds) {
+            SqlStatementBuilder sql = new SqlStatementBuilder();
+            sql.append("DELETE FROM CallReportBase WHERE CallId IN (", callReportIds, ")");
+            return sql.toString();
+        }
     }
     
     @Override
@@ -43,4 +61,8 @@ public class CallReportDaoImpl implements CallReportDao {
         this.ecMappingDao = ecMappingDao;
     }
 
+    @Override
+    public void afterPropertiesSet() throws Exception {
+        chunkyJdbcTemplate= new ChunkingSqlTemplate<Integer>(simpleJdbcTemplate);
+    }
 }
